requirements.txt
backports.entry-points-selectable==1.1.0
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.2
et-xmlfile==1.1.0
ExifRead==2.3.2
filelock==3.0.12
idna==2.10
jdcal==1.4.1
openpyxl==2.5.4
Pillow==8.1.0
platformdirs==2.2.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
six==1.16.0
urllib3==1.26.3
virtualenv==20.7.0
xlrd==1.2.0
xlwt==1.3.0
